refactored circumpoints API

+++ API/serializers.py
@@ -3,4 +3,4 @@ from rest_framework import serializers
 class ConvertCoordinatesSerializer(serializers.Serializer):
     length = serializers.IntegerField()
     width = serializers.IntegerField()
-    value = serializers.IntegerField()
+    value = serializers.FloatField()
